Implement the client specification test
Unfortunately Go doesn't have a way to specify a dependency that doesn't
contain Go files so the only way to fix this at the moment is to clone
the test specification into the repository when Travis builds it. People
developing locally would have to do the same. There is new support for
modules coming soon so perhaps this can be soved another way.

Fixes #26