Fix unit tests by removing global unleash instances
When running the tests, Go will compile all _test.go files into the
binary. When executing the tests all init() functions are called when
the binary is loaded so this example code was creating two instances of
Unleash clients that were messing with tests because they were sending
unexpected HTTP requests which the mocker didn't like.